chore: rename chat_message file
